django-server/requirements.txt
django==4.2.11
djangorestframework==3.15.1
psycopg2-binary==2.9.9
dj-database-url==2.1.0
drf-spectacular==0.27.2